requirements.txt
Django==4.2.11
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-channels
django-celery-beat==2.5.0
django-celery-results==2.5.1
psycopg2-binary==2.9.9
